@@ -12,6 +12,7 @@ const mix = require('laravel-mix');
  */
 
 mix.js('resources/js/app.js', 'public/js')
-    .postCss('resources/css/app.css', 'public/css', [
-        //
-    ]);
+    .postCss('resources/css/app.css', 'public/css');
+
+mix.js('resources/js/dashboard.js', 'public/dashboard/js')
+    .postCss('resources/css/dashboard/dashboard.css', 'public/dashboard/css');
